Add basic code to read and parse the log files.
Add -billion flag and billion() function, but this code doesn't actually
calculate a prediction yet.  It does open the log file, read it, find
the earliest and latest timestamps and associated counts, parses those
into time.Time and int64 objects, and prints the results.  The next step
is to actually calculate the estimated target date.

Co-authored-by: Cassandra Corzine <cassie@ties.org>